net_util: refactor unix for cross platforming.
BUG=b:213151463

TEST=bots

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3598233
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>

+++ devices/src/virtio/net.rs
@@ -582,7 +582,7 @@ where
 // Ensure that the tap interface has the correct flags and sets the offload and VNET header size
 // to the appropriate values.
 pub fn validate_and_configure_tap<T: TapT>(tap: &T, vq_pairs: u16) -> Result<(), NetError> {
-    let flags = tap.if_flags();
+    let flags = tap.if_flags() as i32;
     let mut required_flags = vec![
         (libc::IFF_TAP, "IFF_TAP"),
         (libc::IFF_NO_PI, "IFF_NO_PI"),
